add template report api testing
